# Use an optimised conversion proc for Postgres timestamps