@Gerg 2021-12-13
Does MySQL have the same problem?