@philippthun 2021-12-14
> Does MySQL have the same problem?

@Gerg - short answer: no, long answer: thanks for asking!

So I looked into the topic of timestamp conversion using Sequel and MySQL and couldn't find similar conversions as implemented in Sequel's PostgreSQL adapter. So the conversion is already done "below" in the `mysql2` gem that's using `libmysql` and is already fast. Here a comparison between PostgreSQL and MySQL when fetching a single user record from the database 10000 times:

### PostgreSQL

| user | system | total |
| -- | -- | -- |
|11.120050|18.784816|29.904866|
|11.189934|18.721485|29.911419|

### MySQL

| user | system | total |
| -- | -- | -- |
|0.782401|0.229683|1.012084|
|0.734940|0.210260|0.945200|

When running the same benchmark with the optimisation from this PR, it is slightly faster than MySQL:

### PostgreSQL with optimised conversion proc for timestamps

| user | system | total |
| -- | -- | -- |
|0.685741|0.257086|0.942827|
|0.704644|0.262820|0.967464|

So, no, MySQL does not have this performance problem.

But then I stumbled upon the [PostgreSQL-specific Support in Sequel](https://github.com/jeremyevans/sequel/blob/master/doc/postgresql.rdoc) documentation and learned about `sequel_pg`. By simply adding this gem (and removing my manual optimisation), the benchmark results look as follows:

### PostgreSQL with sequel_pg

| user | system | total |
| -- | -- | -- |
|0.596703|0.256924|0.853627|
|0.591460|0.250780|0.842240|

So I've created another PR #2595 and would suggest to merge that one and close this one.