@Gerg 2021-12-14
@philippthun That's great! Looks like there could be some other benefits as well:

> When the postgres adapter is used with the pg driver, Sequel automatically checks for sequel_pg, and loads it if it is available. sequel_pg is a C extension that optimizes the fetching of rows, generally resulting in a ~2x speedup. It is highly recommended to install sequel_pg if you are using the postgres adapter with pg.

> sequel_pg has additional optimizations when using the Dataset map, as_hash, to_hash_groups, select_hash, select_hash_groups, select_map, and select_order_map methods, which avoids creating intermediate hashes and can add further speedups.